requirements.txt
opencv-python
numpy
matplotlib
Pillow
keras
sklearn
tensorflow
pandas